# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ADMIS API v1",
        "description": "API documentation for ADMIS",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://dev-api.admis.fd.cvut.cz",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user & obtain JWT token",
                "description": "Authenticates a user using their credentials and returns a JWT token. See the User model for details on user fields.",
                "operationId": "bb41d9524d1781b0e0f057673979e5c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "description": "User's unique username",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "The user's password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User successfully authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "JWT access token",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "description": "Token type",
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "description": "Time in seconds until the token expires",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Retrieve details of the authenticated user",
                "description": "Returns the details of the authenticated user. Data is based on the User model. Refer to App\\Models\\Users\\User for field definitions.",
                "operationId": "7029e3e8d2d6c300d23adff07cfb497e",
                "responses": {
                    "200": {
                        "description": "Authenticated user retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "username": {
                                            "description": "User's unique identifier",
                                            "type": "string"
                                        },
                                        "name": {
                                            "description": "Full name of the user",
                                            "type": "string"
                                        },
                                        "email": {
                                            "description": "Email address",
                                            "type": "string",
                                            "format": "email"
                                        },
                                        "idOu": {
                                            "description": "Organizational unit ID",
                                            "type": "integer"
                                        },
                                        "idRoleType": {
                                            "description": "Role type identifier",
                                            "type": "integer"
                                        },
                                        "idAuthorityType": {
                                            "description": "Authority type identifier",
                                            "type": "integer"
                                        },
                                        "accessDenied": {
                                            "description": "Access status",
                                            "type": "boolean"
                                        },
                                        "idReportConfig": {
                                            "description": "Report configuration identifier",
                                            "type": "integer"
                                        },
                                        "editorReport": {
                                            "description": "Flag for report editor rights",
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - token is missing or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user and invalidate JWT token",
                "description": "Logs out the authenticated user and invalidates their JWT token.",
                "operationId": "944f39c800d7ae24b55b18868e032d07",
                "responses": {
                    "200": {
                        "description": "User logged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - token is missing or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh JWT token",
                "description": "Generates a new JWT token using the current valid token, invalidating the old one. See the User model for token claim structure.",
                "operationId": "ffaeee28eb14655973d0c37360e5d73d",
                "responses": {
                    "200": {
                        "description": "JWT token refreshed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "New JWT access token",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "description": "Type of token",
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "description": "Token expiration time in seconds",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - token is missing or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}